chore: Remove Redis compatibility layer from test stubs
